refactor: move `interpolatePanelStrings` to own file
